Rename files to match id_name
